requirements.txt
discord.py-self
azcaptchaapi
requests
datetime
asyncio